Set metadata in authenticator component
